pyproject.toml
[tool.poetry]
name = "amta"
version = "0.1.0"
description = ""
authors = ["Vitalik Svt <vitalik.svt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.4"
pydantic = {extras = ["email"], version = "^2.9.2"}
uvicorn = "^0.32.0"
pymongo = "^4.9"
pytest = "^8.3.3"
itsdangerous = "^2.2.0"
Jinja2 = "^3.1.4"
pydantic-settings = "^2.6.1"
routers = "^0.10.1"
motor = "^3.6.0"
pyyaml = "^6.0.2"
pyjwt = "^2.10.1"
passlib = "^1.7.4"
python-multipart = "^0.0.19"
jsonref = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
